requirements.txt
# Core web app frameworks
# gradio

# Discord bot
discord.py

# Utilities
python-dotenv